for 4 and crxn for 0 and 1

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -4,7 +4,7 @@ import random
 import asyncio
 
 
-async def wait_random(max_delay: float = 10) -> float:
+async def wait_random(max_delay: int = 10) -> float:
     """ wait random time and return delay """
     delay: float = random.uniform(0, max_delay)
     await asyncio.sleep(delay)
